A bit of work
Added in addNewRoot. This checks if a desired root directory is already
referenced by the database in some way (including as a subdirectory of a
directory that already is in there.) We then add the subdirectory and
get its key value. We get a list of all its subdirectories, remove any
that are already covered by previously added root directories, and end
up with a list of new subdirectories, relative to the new root
directory.

Also, I've added a few directories to be used in live testing.