It's WORKING!!!!
It's glitchy and buggy as heck, but display.py is working and displays pictures based on the database. :)!! Writing in python because a) I need to work on my python skills and b) I've done inter-process communication in python before, and it helps to have that experience for the future of the project.